+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fixed bug in `Update-FabricSemanticModelDefinition` - Uri was incorrect when a platform file exists
 - Name of new SQL database will be shown (#169)
+- Logging in with a Service Principal does not work (#164)
 
 ### Deprecated